requirements.txt
streamlit
opencv-python-headless
ultralytics
streamlit-webrtc
av
Pillow